+++ app/components/primer/markdown.rb
@@ -4,6 +4,8 @@ module Primer
   # Use `Markdown` to wrap markdown content
   class Markdown < Primer::Component
     status :beta
+
+    DEFAULT_TAG = :div
     # @example Default
     #   <%= render(Primer::Markdown.new) do %>
     #     <p>Text can be <b>bold</b>, <i>italic</i>, or <s>strikethrough</s>. <a href="https://github.com">Links </a> should be blue with no underlines (unless hovered over).</p>
@@ -276,10 +278,11 @@ module Primer
     #     <pre><code>This is the final element on the page and there should be no margin below this.</code></pre>
     #   <% end %>
     #
+    # @param tag [Symbol]
     # @param system_arguments [Hash] <%= link_to_system_arguments_docs %>
-    def initialize(**system_arguments)
+    def initialize(tag: DEFAULT_TAG, **system_arguments)
       @system_arguments = system_arguments
-      @system_arguments[:tag] ||= :div
+      @system_arguments[:tag] ||= tag
 
       @system_arguments[:classes] = class_names(
         "markdown-body",