+++ test/components/octicon_component_test.rb
@@ -49,6 +49,18 @@ class PrimerOcticonComponentTest < Minitest::Test
     assert_selector("[style='margin-left: 100px']")
   end
 
+  def test_renders_size_small_when_height_is_less_than_small
+    render_inline(Primer::OcticonComponent.new(:star, height: 12))
+
+    assert_selector("[height='16']")
+  end
+
+  def test_renders_size_small_when_width_is_less_than_small
+    render_inline(Primer::OcticonComponent.new(:star, width: 12))
+
+    assert_selector("[height='16']")
+  end
+
   def test_renders_default_size_small
     render_inline(Primer::OcticonComponent.new(:star))