+++ CHANGELOG.md
@@ -21,6 +21,9 @@
 * Do not raise error if Primer CSS class name is passed to component if `PRIMER_WARNINGS_DISABLED` is set.
 
     *Joel Hawksley*
+### Breaking changes
+
+* Restrict allowed tags for `Truncate` and `HiddenTextExpander` components.
 
 ### Accessibility